optee-teec: support multiple TEE implementations
TODO: also need to abstract parameter construction (TEEC_PARAM_TYPES
macro in C), since that's implementation-defined too.

The GlobalPlatform TEE Client API this crate binds to comes very close
to defining a single ABI for all implementations, but it doesn't quite
get there. It puts the client in charge of allocating API objects like
TEEC_Session, but it allows implementations to add an arbitrary amount
of data to the end of several of those objects, meaning their sizes
differ per implementation.

Since everything else is generic, those fields are the only part of this
crate that's specific to OP-TEE. This commit moves those field
definitions into a new set of `_imp` structures and selects between
multiple versions of those structures based on a new set of feature
flags.

An added plus of this is that the implementation-specific fields are
hidden to users, as per the spec's requirement that "Client Application
code MUST NOT directly access the contents of the `imp` field." We also
eliminate a use of the dangerous mem::zeroed(), since all the new imp
structs are wrapped in a MaybeUninit, allowing us to instantiate them
without seeing their fields.