WIP: Move common package fields to workspace
PROBLEM: Currently blocked on the fact that the nightly-2021-09-20 we
use is from before Cargo 1.64 and so doesn't have this feature.

Cargo 1.64 added the ability for crates in a workspace to inherit
settings from Cargo.toml in the workspace root. Make use of it to give
us a single source of truth for version, license, URL, and author.